extract chess-related types
