+++ src/chess/mod.rs
@@ -1,3 +1,4 @@
 pub mod board;
 pub mod chessmove;
 pub mod piece;
+pub mod types;
