+++ src/threadlocal.rs
@@ -3,12 +3,12 @@ use std::array;
 use crate::{
     chess::board::Board,
     chess::chessmove::Move,
+    chess::piece::Colour,
     historytable::{
         CaptureHistoryTable, CorrectionHistoryTable, DoubleHistoryTable, MoveTable,
         ThreatsHistoryTable,
     },
     nnue::{self, network::NNUEParams},
-    chess::piece::Colour,
     search::pv::PVariation,
     stack::StackEntry,
     transpositiontable::TTView,
